+++ roqoqo/src/measurements/mod.rs
@@ -45,16 +45,16 @@ use crate::{
     Circuit, RoqoqoError,
 };
 
-#[cfg(feature="async")]
-use futures::{ future::FutureExt};
-#[cfg(feature="async")]
-use async_trait::async_trait;
-#[cfg(feature="async")]
-use std::pin::Pin;
-#[cfg(feature="async")]
-use crate::RoqoqoBackendError;
-
+#[cfg(feature = "async")]
 use crate::registers::Registers;
+#[cfg(feature = "async")]
+use crate::RoqoqoBackendError;
+#[cfg(feature = "async")]
+use async_trait::async_trait;
+#[cfg(feature = "async")]
+use futures::future::FutureExt;
+#[cfg(feature = "async")]
+use std::pin::Pin;
 
 /// Allows generic interfacing with roqoqo measurements.
 ///
@@ -194,7 +194,7 @@ pub trait Measure: PartialEq + Clone {
 /// assert_eq!(result.get("single_qubit_exp_val").unwrap(), &0.0);
 /// ```
 ///
-#[cfg_attr(feature="async", async_trait)]
+#[cfg_attr(feature = "async", async_trait)]
 pub trait MeasureExpectationValues: PartialEq + Clone + Measure {
     /// Evaluates measurement results based on classical registers.
     ///
@@ -227,12 +227,16 @@ pub trait MeasureExpectationValues: PartialEq + Clone + Measure {
     /// * `Ok(Some(HashMap<String, f64>))` - The measurement has been evaluated successfully. The HashMap contains the measured expectation values.
     /// * `Ok(None)` - The measurement did not fail but is incomplete. A new round of measurements is needed.
     /// * `Err(RoqoqoError)` - The measurement evaluation failed.
-    #[cfg(feature="async")]
+    #[cfg(feature = "async")]
     async fn async_evaluate(
         &self,
-        registers: Pin<Box<dyn FutureExt<Output=Result<Registers, RoqoqoBackendError>> + std::marker::Send >>
-    ) -> Result<HashMap<String, f64>, RoqoqoBackendError>{
+        registers: Pin<
+            Box<dyn FutureExt<Output = Result<Registers, RoqoqoBackendError>> + std::marker::Send>,
+        >,
+    ) -> Result<HashMap<String, f64>, RoqoqoBackendError> {
         let (bit_registers, float_registers, complex_registers) = registers.await?;
-        Ok(self.evaluate(bit_registers, float_registers, complex_registers)?.unwrap())
+        Ok(self
+            .evaluate(bit_registers, float_registers, complex_registers)?
+            .unwrap())
     }
 }