+++ roqoqo/src/quantum_program.rs
@@ -14,12 +14,12 @@
 
 use std::collections::HashMap;
 
+#[cfg(feature = "async")]
+use crate::backends::AsyncEvaluatingBackend;
 use crate::backends::{EvaluatingBackend, RegisterResult};
 use crate::measurements;
 use crate::measurements::Measure;
 use crate::RoqoqoBackendError;
-#[cfg(feature="async")]
-use crate::backends::AsyncEvaluatingBackend;
 use std::fmt::{Display, Formatter};
 
 /// Represents a quantum program evaluating measurements based on a one or more free float parameters.
@@ -145,7 +145,6 @@ impl QuantumProgram {
         }
     }
 
-
     /// Runs the QuantumProgram and returns expectation values.
     ///
     /// Runs the quantum programm for a given set of parameters passed in the same order as the parameters
@@ -155,7 +154,7 @@ impl QuantumProgram {
     ///
     /// * `backend` - The backend the program is executed on.
     /// * `parameters` - List of float ([f64]) parameters of the function call in order of `input_parameter_names`
-    #[cfg(feature="async")]
+    #[cfg(feature = "async")]
     pub async fn async_run<T>(
         &self,
         backend: T,
@@ -163,7 +162,7 @@ impl QuantumProgram {
     ) -> Result<Option<HashMap<String, f64>>, RoqoqoBackendError>
     where
         T: AsyncEvaluatingBackend,
-        T: Sync
+        T: Sync,
     {
         match self{
             QuantumProgram::PauliZProduct{measurement, input_parameter_names } => {
@@ -206,11 +205,11 @@ impl QuantumProgram {
     ///
     /// * `backend` - The backend the program is executed on.
     /// * `parameters` - List of float ([f64]) parameters of the function call in order of `input_parameter_names`
-    #[cfg(feature="async")]
+    #[cfg(feature = "async")]
     pub async fn async_run_registers<T>(&self, backend: T, parameters: &[f64]) -> RegisterResult
     where
         T: AsyncEvaluatingBackend,
-        T: Sync
+        T: Sync,
     {
         match self{
             QuantumProgram::ClassicalRegister{measurement, input_parameter_names } => {