+++ roqoqo/tests/integration/circuit.rs
@@ -193,6 +193,7 @@ fn remap_qbits() {
 
     let mut qubit_mapping_test: HashMap<usize, usize> = HashMap::new();
     qubit_mapping_test.insert(2, 0);
+    qubit_mapping_test.insert(0, 2);
     let result = circuit_test.remap_qubits(&qubit_mapping_test).unwrap();
     assert_eq!(result, circuit)
 }